requirements.txt
beautifulsoup4
pandas
nltk
requests-html
flask
firebase-admin
google-cloud-bigquery
pyarrow
